requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
SQLAlchemy==1.3.19
alembic==1.4.3
psycopg2-binary==2.8.6
PyJWT==1.7
bcrypt==3.2.0
validate_email==1.3
pytz==2021.1
pytest==6.2.3
requests==2.25.1
flake8==3.9.0